:microscope: improve test coverage

+++ tests/test_main.py
@@ -1,6 +1,9 @@
 import os
+import sys
 from shutil import copyfile
 from nose.tools import raises
+from mock import patch
+import moban.exceptions as exceptions
 
 
 class TestException:
@@ -9,7 +12,8 @@ class TestException:
         self.data_file = 'data.yml'
 
     def tearDown(self):
-        os.unlink(self.moban_file)
+        if os.path.exists(self.moban_file):
+            os.unlink(self.moban_file)
         if os.path.exists(self.data_file):
             os.unlink(self.data_file)
         if os.path.exists('.moban.hashes'):
@@ -30,3 +34,27 @@ class TestException:
                  self.data_file)
         import moban.main as main
         main.handle_moban_file({})
+
+    @patch('os.path.exists')
+    @patch('moban.main.handle_moban_file')
+    @patch('moban.reporter.report_error_message')
+    def test_directory_not_found(self,
+                                 fake_reporter, fake_moban_file, fake_file):
+        fake_file.return_value = True
+        fake_moban_file.side_effect = exceptions.DirectoryNotFound
+        from moban.main import main
+        with patch.object(sys, 'argv', ['moban']):
+            main()
+        assert fake_reporter.called
+
+    @patch('os.path.exists')
+    @patch('moban.main.handle_moban_file')
+    @patch('moban.reporter.report_error_message')
+    def test_no_third_party_engine(self,
+                                   fake_reporter, fake_moban_file, fake_file):
+        fake_file.return_value = True
+        fake_moban_file.side_effect = exceptions.NoThirdPartyEngine
+        from moban.main import main
+        with patch.object(sys, 'argv', ['moban']):
+            main()
+        assert fake_reporter.called
